+++ GenLauncherNet/GenLauncher.csproj
@@ -148,7 +148,7 @@
     <Compile Include="DataClasses\ModificationFileInfo.cs" />
     <Compile Include="DataClasses\StringHashSet.cs" />
     <Compile Include="SessionInformation.cs" />
-    <Compile Include="Utilities.cs" />
+    <Compile Include="Utility\Utilities.cs" />
     <Compile Include="WPFElements\UpdateButton.cs" />
     <Compile Include="Windows\Net46NotInstalled.xaml.cs">
       <DependentUpon>Net46NotInstalled.xaml</DependentUpon>
@@ -354,6 +354,7 @@
   <ItemGroup>
     <Resource Include="Images\BackgroundGenerals.png" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\SevenZipExtractor.1.0.15\build\SevenZipExtractor.targets" Condition="Exists('..\packages\SevenZipExtractor.1.0.15\build\SevenZipExtractor.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
